requirements.txt
google-adk
uuid
